+++ pom.xml
@@ -9,22 +9,27 @@
     <version>0.1</version>
     <name>KonfiguracjaSpringaAdnotacjami</name>
     <description>KonfiguracjaSpringaAdnotacjami</description>
+    <packaging>war</packaging>
 
     <properties>
-
         <!-- Generic properties -->
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+
+        <!--  Maven -->
+        <maven.compiler.version>3.1</maven.compiler.version>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
 
         <!-- Web -->
         <servlet.version>3.0.1</servlet.version>
         <jstl.version>1.2</jstl.version>
+        <tomcat.version>2.2</tomcat.version>
 
         <!-- Spring -->
         <spring-bom.version>5.1.7.RELEASE</spring-bom.version>
         <jackson.version>1.9.13</jackson.version>
-
     </properties>
 
     <dependencyManagement>
@@ -90,20 +95,17 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.1</version>
+                <version>${maven.compiler.version}</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
                     <showDeprecation>true</showDeprecation>
                     <showWarnings>true</showWarnings>
                 </configuration>
             </plugin>
             <plugin>
-                <groupId>org.eclipse.jetty</groupId>
-                <artifactId>jetty-maven-plugin</artifactId>
-                <version>9.2.0.v20140526</version>
+                <groupId>org.apache.tomcat.maven</groupId>
+                <artifactId>tomcat7-maven-plugin</artifactId>
+                <version>${tomcat.version}</version>
             </plugin>
         </plugins>
     </build>
 </project>
-
