--- src/main/java/a_adnotacje_wstrzykiwania/autowired/komponenty_klawiatury/KomponentKlawiatury.java
@@ -1,12 +1,0 @@
-package a_adnotacje_wstrzykiwania.autowired.komponenty_klawiatury;
-
-/**
- * Interfejs, który reprezentuje komponent (składnik) klawiatury.
- *
- * @author Wiktor Rup
- */
-public interface KomponentKlawiatury {
-
-    void robiRzeczy();
-
-}
