--- src/main/java/a_adnotacje_wstrzykiwania/autowired/komponenty_klawiatury/Przycisk.java
@@ -1,15 +1,0 @@
-package a_adnotacje_wstrzykiwania.autowired.komponenty_klawiatury;
-
-/**
- * Reprezentacja Przycisku, czyli zależność, która zostaje wstrzyknięta do klasy {@link a_adnotacje_wstrzykiwania.autowired.komponenty_komputera.Klawiatura Klawiatura} przez pole.
- *
- * @author Wiktor Rup
- * @see KomponentKlawiatury
- */
-class Przycisk implements KomponentKlawiatury {
-
-    @Override
-    public void robiRzeczy() {
-        System.out.println("Klik...");
-    }
-}
