--- src/main/java/a_adnotacje_wstrzykiwania/autowired/komponenty_komputera/IPort.java
@@ -1,11 +1,0 @@
-package a_adnotacje_wstrzykiwania.autowired.komponenty_komputera;
-
-/**
- * Interfejs, który jest ogólną reprezentacją różnego rodzaju portów.
- *
- * @author Wiktor Rup
- */
-public interface IPort {
-
-    void status();
-}
