--- src/main/java/a_adnotacje_wstrzykiwania/autowired/komponenty_komputera/KomponentKomputera.java
@@ -1,11 +1,0 @@
-package a_adnotacje_wstrzykiwania.autowired.komponenty_komputera;
-
-/**
- * Interfejs, który reprezentuje komponent (składnik) komputera.
- *
- * @author Wiktor Rup
- */
-public interface KomponentKomputera {
-
-    void status();
-}
