--- src/main/java/a_adnotacje_wstrzykiwania/qualifier/Elektrownia.java
@@ -1,33 +1,0 @@
-package a_adnotacje_wstrzykiwania.qualifier;
-
-import a_adnotacje_wstrzykiwania.qualifier.komponenty_elektrowni.Reaktor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-
-/**
- * Klasa prezentująca działanie adnotacji Qualifier. {@link org.springframework.beans.factory.annotation.Qualifier}
- *
- * @author Wiktor Rup
- */
-class Elektrownia {
-
-
-    private Reaktor reaktor;
-
-    /**
-     * Wstrzyknięcie zależności za pomocą settera. Koniecznym jest zastosowanie adnotacji Qualifier aby określić, która fasolka ma zostać wstrzyknięta.
-     * Interfejs Reaktor implementowany jest zarówno przez klasę ReaktorChemiczny jak i ReaktorJądrowy, dlatego należy zdefiniować konkretnie, która fasolka nas interesuje.
-     *
-     * @param reaktor
-     * @see Reaktor
-     */
-    @Autowired
-    @Qualifier("reaktorchemiczny")
-    public void setReaktor(Reaktor reaktor) {
-        this.reaktor = reaktor;
-    }
-
-    void rodzajElektrowni() {
-        reaktor.typ();
-    }
-}
