--- src/main/java/a_adnotacje_wstrzykiwania/qualifier/komponenty_elektrowni/Reaktor.java
@@ -1,10 +1,0 @@
-package a_adnotacje_wstrzykiwania.qualifier.komponenty_elektrowni;
-
-/**
- * Interfejs który jest reprezentacją reaktora stosowanego w elektrownii.
- *
- * @author Wiktor Rup
- */
-public interface Reaktor {
-    void typ();
-}
