--- src/main/java/a_adnotacje_wstrzykiwania/required/package-info.java
@@ -1,16 +1,0 @@
-/**
- * <h1>Adnotacja Required!</h1>
- *
- *  <p>Podczas zapoznania się z kodem w tym pakiecie dowiesz się o adnotacji {@link org.springframework.beans.factory.annotation.Required Required},
- *  która wymusza na użytkowniku wstrzyknięcie zależności przy inicjalizacji danej klasy.</p>
- *  <br>
- *
- *  <p>Przeanalizuj klasę {@link a_adnotacje_wstrzykiwania.required.Samochód Samochód} w pakiecie <font color="yellow">{@code required}</font>.</p>
- *
- *  <p>Jednocześnie w katalogu <font color="yellow">{@code resources}</font> znajdziesz plik konfiguracyjny dla tego pakietu o nazwie {@code RequiredBeans.xml}.
- *  Spróbuj go zmodyfikować tak aby zauważyć działanie adnotacji Required.</p>
- *
- *
- * @author Wiktor Rup
- */
-package a_adnotacje_wstrzykiwania.required;
