--- src/main/java/b_przykład_konfiguracji_adnotacjami/OszołomionaOsoba.java
@@ -1,17 +1,0 @@
-package b_przykład_konfiguracji_adnotacjami;
-
-/**
- * Proste POJO. Ot zwykła otoczka na imię.
- *
- * @author Wojciech Makiela
- */
-class OszołomionaOsoba {
-
-    String imie;
-
-    OszołomionaOsoba(String imie) {
-        this.imie = imie;
-        System.out.println(this.getClass().getName() + " - Konstruktor");
-
-    }
-}
