+++ src/main/java/ja/workshops/spring/części/konfiguracja/adnotacjami/a_adnotacje_wstrzykiwania/autowired/komponenty_klawiatury/KomponentKlawiatury.java
@@ -1,0 +1,12 @@
+package ja.workshops.spring.części.konfiguracja.adnotacjami.a_adnotacje_wstrzykiwania.autowired.komponenty_klawiatury;
+
+/**
+ * Interfejs, który reprezentuje komponent (składnik) klawiatury.
+ *
+ * @author Wiktor Rup
+ */
+public interface KomponentKlawiatury {
+
+    void robiRzeczy();
+
+}
