+++ src/main/java/ja/workshops/spring/części/konfiguracja/adnotacjami/a_adnotacje_wstrzykiwania/autowired/komponenty_komputera/IPort.java
@@ -1,0 +1,11 @@
+package ja.workshops.spring.części.konfiguracja.adnotacjami.a_adnotacje_wstrzykiwania.autowired.komponenty_komputera;
+
+/**
+ * Interfejs, który jest ogólną reprezentacją różnego rodzaju portów.
+ *
+ * @author Wiktor Rup
+ */
+public interface IPort {
+
+    void status();
+}
