+++ src/main/java/ja/workshops/spring/części/konfiguracja/adnotacjami/a_adnotacje_wstrzykiwania/autowired/komponenty_komputera/KomponentKomputera.java
@@ -1,0 +1,11 @@
+package ja.workshops.spring.części.konfiguracja.adnotacjami.a_adnotacje_wstrzykiwania.autowired.komponenty_komputera;
+
+/**
+ * Interfejs, który reprezentuje komponent (składnik) komputera.
+ *
+ * @author Wiktor Rup
+ */
+public interface KomponentKomputera {
+
+    void status();
+}
