+++ src/main/java/ja/workshops/spring/części/konfiguracja/adnotacjami/a_adnotacje_wstrzykiwania/qualifier/komponenty_elektrowni/Reaktor.java
@@ -1,0 +1,10 @@
+package ja.workshops.spring.części.konfiguracja.adnotacjami.a_adnotacje_wstrzykiwania.qualifier.komponenty_elektrowni;
+
+/**
+ * Interfejs który jest reprezentacją reaktora stosowanego w elektrownii.
+ *
+ * @author Wiktor Rup
+ */
+public interface Reaktor {
+    void typ();
+}
