+++ src/main/resources/RequiredBeans.xml
@@ -8,7 +8,7 @@
          http://www.springframework.org/schema/context/spring-context-3.0.xsd">
     <context:annotation-config/>
     <!-- Zdefiniowanie ziarna samochód wraz z wymaganymi argumentami -->
-    <bean id = "samochód" class = "a_adnotacje_wstrzykiwania.required.Samochód">
+    <bean id = "samochód" class = "ja.workshops.spring.części.konfiguracja.adnotacjami.a_adnotacje_wstrzykiwania.required.Samochód">
         <property name = "rokProdukcji" value = "2016" />
         <property name = "marka" value = "Opel"/>
     </bean>
