+++ src/main/resources/spring/mvc-context.xml
@@ -7,5 +7,5 @@
            http://www.springframework.org/schema/context
            http://www.springframework.org/schema/context/spring-context-3.0.xsd">
 
-  <context:component-scan base-package="b_przykład_konfiguracji_adnotacjami" />
+  <context:component-scan base-package="ja.workshops.spring.części.konfiguracja.adnotacjami.b_przykład_konfiguracji" />
 </beans>
